CAI_NetworkBuilder::SaveNetworkGraph improvements
Use CIOStream class.
Use proper type for CAI_Network::m_iNumNodes.
Improved CIOStream to take std::filesystem::path instead of std::string.

+++ r5dev/game/server/ai_network.h
@@ -40,7 +40,7 @@ public:
 
 	char pad0[0x14];              // +0x1054 <-- !TODO
 
-	int64_t m_iNumNodes;          // +0x1070
+	int m_iNumNodes;              // +0x1070
 	CAI_Node** m_pAInode;         // +0x1078
 };
